+++ src/connection/socket.py
@@ -26,26 +26,56 @@ class Socket(QObject):
         self.status_widget.set_status_text("Client Connected Event")
 
     def on_disconnected(self):
-        LOGGER.debug('Client disconnect event')
+        LOGGER.debug('-*- Client disconnect event -*-')
         self.status_widget.set_status_text("Client socket closed.")
 
-    def on_readyRead(self):
-        LOGGER.debug('Socket message ready')
+    def _get_message(self):
+        """Get the socket message.
 
+        If the message is a simple string then return it into a dictionary,
+        else return deserialized array.
+
+        Will raise a ValueError exception if can not convert the stringified array.
+
+        Returns:
+            dict - dictionary data with the following keys:
+                'text': code to execute
+                'file': optional file name
+        """
         msg = self.socket.readAll()
 
+        # startsWith is a method of QByteArray
+        if not msg.startsWith('{'):
+            LOGGER.debug('Message is simple string. Converting into dict')
+            return {'text': msg.data()}
+
         try:
+            LOGGER.debug('Message is stringified array.')
             msg_data = json.loads(msg.data())
         except Exception as err:
             LOGGER.critical("Invalid Json Deserialization: %s",
                             err, exc_info=True)
+            raise ValueError('Json Deserialization error')
+        else:
+            LOGGER.debug('msg_data: %s', msg_data)
+            return msg_data
+
+    def on_readyRead(self):
+        LOGGER.debug('Socket message ready')
+
+        try:
+            msg_data = self._get_message()
+        except ValueError:
             return
 
-        msg_to_string = msg_data['text']
-        LOGGER.debug('msg data type: %s', type(msg))
+        msg_text = msg_data.get('text')
+        if not msg_text:
+            LOGGER.warning("no text data to execute")
+            return
 
-        script_editor = ScriptEditor(file=msg_data.get('file', ''))
-        script_editor.set_text(msg_to_string)
+        script_editor = ScriptEditor()
+        script_editor.set_file(msg_data.get('file', ''))
+        script_editor.set_text(msg_text)
         script_editor.execute()
 
         output_text = script_editor.set_status_output()
@@ -58,5 +88,5 @@ class Socket(QObject):
         self.socket.close()
         LOGGER.debug('Closing socket')
 
-        self.status_widget.set_input_text(msg_to_string)
+        self.status_widget.set_input_text(msg_text)
         self.status_widget.set_output_text(output_text)