+++ src/widgets/script_editor.py
@@ -1,4 +1,3 @@
-"""Fake class imitating the internal Nuke Script Editor needed for testing."""
 # coding: utf-8
 from __future__ import print_function
 
@@ -6,9 +5,12 @@ import sys
 import random
 import subprocess
 
-from PySide2.QtCore import Qt
+
+from PySide2.QtCore import Qt, QEvent
+from PySide2.QtGui import QKeySequence, QKeyEvent
 
 from PySide2.QtWidgets import (
+    QShortcut,
     QVBoxLayout,
     QWidget,
     QPlainTextEdit,
@@ -29,8 +31,7 @@ class InputEditor(QPlainTextEdit):
     def __init__(self):
         QPlainTextEdit.__init__(self)
         r = random.randint(1, 50)
-        self.setPlainText(
-            "import json; data=json.loads('{\"text\":\"Hello from Test Client\", \"num\": %s }'); print data['text'], data['num']" % r)
+        self.setPlainText("print('Hello From Internal SE %s')" % r)
 
 
 class FakeScriptEditor(QWidget):
@@ -40,9 +41,11 @@ class FakeScriptEditor(QWidget):
 
         self.run_btn = QPushButton('Run')
         self.run_btn.setToolTip('Run the current')
+        self.run_btn.setShortcut(QKeySequence(Qt.CTRL + Qt.Key_Return))
         self.run_btn.clicked.connect(self.run_code)
 
         self.input_console = InputEditor()
+
         self.output_console = OutputEditor()
 
         self.splitter = QSplitter()
@@ -55,11 +58,21 @@ class FakeScriptEditor(QWidget):
         _layout.addWidget(self.splitter)
         self.setLayout(_layout)
 
+        self.installEventFilter(self)
+
+    def eventFilter(self, obj, event):
+        if isinstance(event, QKeyEvent):
+            if event.modifiers() == Qt.CTRL and event.key() == Qt.Key_Return:
+                print('shortcut pressed')
+                self.run_code()
+                event.accept()
+                return True
+        return super(FakeScriptEditor, self).eventFilter(obj, event)
+
     def run_code(self):
         code = self.input_console.toPlainText()
         call = subprocess.check_output(['python', '-c', code])
         self.output_console.setPlainText(call)
-        # self.output_console.insertPlainText(call)
 
 
 class MainWindow(QMainWindow):